Update Attachments List components
