+++ app/src/features/surveys/view/SurveyAttachments.tsx
@@ -3,67 +3,39 @@ import Divider from '@material-ui/core/Divider';
 import Paper from '@material-ui/core/Paper';
 import { mdiTrayArrowUp } from '@mdi/js';
 import Icon from '@mdi/react';
-import AttachmentsList from 'components/attachments/AttachmentsList';
 import { IReportMetaForm } from 'components/attachments/ReportMetaForm';
 import FileUploadWithMetaDialog from 'components/dialog/attachments/FileUploadWithMetaDialog';
 import { IUploadHandler } from 'components/file-upload/FileUploadItem';
 import { H2MenuToolbar } from 'components/toolbar/ActionToolbars';
 import { SurveyContext } from 'contexts/surveyContext';
 import { useBiohubApi } from 'hooks/useBioHubApi';
-import { IGetSurveyAttachment, IGetSurveyReportAttachment } from 'interfaces/useSurveyApi.interface';
-import React, { useCallback, useContext, useEffect, useState } from 'react';
+import React, { useContext, useState } from 'react';
 import { AttachmentType } from '../../../constants/attachments';
+import SurveyAttachmentsList from './SurveyAttachmentsList';
 
-/**
- * Survey attachments content.
- *
- * @return {*}
- */
 const SurveyAttachments: React.FC = () => {
   const biohubApi = useBiohubApi();
+
   const surveyContext = useContext(SurveyContext);
+
   const { projectId, surveyId } = surveyContext;
 
   const [openUploadAttachments, setOpenUploadAttachments] = useState(false);
   const [attachmentType, setAttachmentType] = useState<AttachmentType.REPORT | AttachmentType.OTHER>(
     AttachmentType.OTHER
   );
-  const [attachmentsList, setAttachmentsList] = useState<IGetSurveyAttachment[]>([]);
-  const [reportAttachmentsList, setReportAttachmentsList] = useState<IGetSurveyReportAttachment[]>([]);
 
-  // Tracks which attachment rows have been selected, via the table checkboxes.
+  surveyContext.artifactDataLoader.load(projectId, surveyId);
 
   const handleUploadReportClick = () => {
     setAttachmentType(AttachmentType.REPORT);
     setOpenUploadAttachments(true);
   };
+
   const handleUploadAttachmentClick = () => {
     setAttachmentType(AttachmentType.OTHER);
     setOpenUploadAttachments(true);
   };
-
-  const getAttachments = useCallback(
-    async (forceFetch: boolean): Promise<IGetSurveyAttachment[] | undefined> => {
-      if (attachmentsList.length && !forceFetch) {
-        return;
-      }
-
-      try {
-        const response = await biohubApi.survey.getSurveyAttachments(projectId, surveyId);
-
-        if (!response?.attachmentsList) {
-          return;
-        }
-
-        setReportAttachmentsList([...response.reportAttachmentsList]);
-        setAttachmentsList([...response.attachmentsList]);
-        return [...response.reportAttachmentsList, ...response.attachmentsList];
-      } catch (error) {
-        return;
-      }
-    },
-    [biohubApi.survey, projectId, surveyId, attachmentsList.length]
-  );
 
   const getUploadHandler = (): IUploadHandler => {
     return (file, cancelToken, handleFileUploadProgress) => {
@@ -81,12 +53,6 @@ const SurveyAttachments: React.FC = () => {
     };
   };
 
-  useEffect(() => {
-    getAttachments(false);
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
   return (
     <>
       <FileUploadWithMetaDialog
@@ -96,8 +62,6 @@ const SurveyAttachments: React.FC = () => {
         onFinish={getFinishHandler()}
         onClose={() => {
           setOpenUploadAttachments(false);
-          getAttachments(true);
-          console.log('getAttachments');
           surveyContext.artifactDataLoader.refresh(projectId, surveyId);
         }}
         uploadHandler={getUploadHandler()}
@@ -116,13 +80,7 @@ const SurveyAttachments: React.FC = () => {
         />
         <Divider></Divider>
         <Box px={1}>
-          <AttachmentsList
-            projectId={projectId}
-            surveyId={surveyId}
-            attachmentsList={[...attachmentsList, ...reportAttachmentsList]}
-            selectedAttachments={[]}
-            getAttachments={getAttachments}
-          />
+          <SurveyAttachmentsList />
         </Box>
       </Paper>
     </>
